+++ vk_raytracing/common/scene_loader.h
@@ -57,7 +57,7 @@ namespace SceneLoader {
 
 		glm::vec3 camera_position = glm::vec3(0, 0, 0);
 		glm::vec3 camera_lookat = glm::vec3(0, 0, 2);
-		unsigned int camera_fov = 45;
+		float camera_fov = 45;
 
 		//Models in scene
 		std::vector<Entity*> entities;
